add DIP-13 and ifx tests

+++ contracts/Migrations.sol
@@ -2,7 +2,7 @@ pragma solidity >=0.4.21 <0.7.0;
 
 contract Migrations {
     address public owner;
-    uint256 public last_completed_migration;
+    uint public last_completed_migration;
 
     constructor() public {
         owner = msg.sender;
@@ -12,7 +12,7 @@ contract Migrations {
         if (msg.sender == owner) _;
     }
 
-    function setCompleted(uint256 completed) public restricted {
+    function setCompleted(uint completed) public restricted {
         last_completed_migration = completed;
     }
